+++ src/views/PageGalleryFilter.svelte
@@ -5,9 +5,14 @@
   const id = ulid()
 
   export let filter: Writable<string>
+
+  function clearFilter () {
+    filter.set('')
+  }
 </script>
 
 <div class="page-gallery__filter">
   <label for={id}>🔎</label>
   <input id={id} type="text" bind:value={$filter}>
+  <a class="page-gallery__filter-clear" title="Clear" href="javascript:;" on:click={clearFilter}>❌</a>
 </div>